refactor: LottoMachine 당첨금액, 보너스번호 설정 LottoCalculator로 역할 위임
