+++ docs/REQUIREMENTS.md
@@ -15,6 +15,14 @@
 - [x] 로또 티켓은 6개의 번호를 가진다.
 - [x] 올바른 로또 번호가 아닐 경우 에러를 발생한다.
 - [x] 로또 금액은 1,000원이다.
+
+로또 계산기
+- 당첨 번호 설정
+  - [x] 당첨 번호를 설정할 수 있다.
+  - [x] 보너스 번호를 설정할 수 있다.
+- 당첨 금액 설정
+  - [x] 당첨 금액은 빈 배열일 수 없다.
+  - [x] 당첨 금액은 길이가 6인 배열이어야 한다.
 - 당첨 금액 조회
   - [x] 1등: 2,000,000,000원
   - [x] 2등: 30,000,000원
@@ -29,18 +37,12 @@
   - [x] 4등: 4개 번호 일치
   - [x] 5등: 3개 번호 일치
   - [x] 낙첨: 2개 이하 번호 일치
+- [x] n개의 로또의 결과를 반환한다
 
-로또 머신
-- 당첨 번호 설정
-  - [x] 당첨 번호를 설정할 수 있다.
-  - [x] 보너스 번호를 설정할 수 있다.
-- 당첨 금액 설정
-  - [x] 당첨 금액은 빈 배열일 수 없다.
-  - [x] 당첨 금액은 길이가 6인 배열이어야 한다.
+로또 기계
 - 로또 발권 (자동)
   - [x] 지불할 금액이 로또 금액보다 적다면 에러를 발생한다.
   - [x] 지불한 금액에 따라 n개의 로또를 발권한다.
-- [x] n개의 로또의 결과를 반환한다
  
 로또 (유틸)
 - [x] 번호는 1~45 사이의 숫자값이다.
